optimize: move some util functions to core and diagrams util
