+++ packages/pintora-core/src/util/index.ts
@@ -5,3 +5,4 @@ export { IFont, calculateTextDimensions } from './text-metric'
 export { encodeForUrl, decodeCodeInUrl } from './encode'
 export { makeMark } from './mark'
 export { parseColor, tinycolor } from './color'
+export { makeArtist } from './artist'
