+++ packages/test-shared/example-files/gantt-1.pintora
@@ -1,0 +1,22 @@
+gantt
+  title Gantt example
+
+  dateFormat YYYY-MM-DDTHH
+  axisFormat MM-DD
+  axisInterval 1w
+
+  section Develop Prototype
+  "Write grammar"       : t-a, 2022-2-17, 2022-2-23
+  "Write artist"        : t-b, 2022-2-23, 2022-3-15
+
+  %% the day I started typing the docs
+  markDate 2022-3-15T20
+
+  section Documentation
+  "Write docs"          : t-c, 2022-3-15, 5d
+
+  section Optimize
+  "Add axisInterval" : 2022-3-28, 2022-4-04
+
+  section Release
+  "Release" : milestone, 2022-4-06, 0d
